requirements.txt
Pillow==9.4.0
python-dotenv==0.21.1
